requirement.txt
streamlit
pymupdf
langchain
langchain-ollama
faiss-cpu
numpy
pandas
nltk
spacy